Updated Migration EventHandler and PendingMigrationChecker
