+++ shared/EShopOnAbp.Shared.Hosting.Microservices/DbMigrations/EfCore/PendingEfCoreMigrationsChecker.cs
@@ -1,0 +1,67 @@
+﻿using System;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Volo.Abp.Data;
+using Volo.Abp.EventBus.Distributed;
+using Volo.Abp.MultiTenancy;
+using Volo.Abp.Uow;
+
+namespace EShopOnAbp.Shared.Hosting.Microservices.DbMigrations.EfCore;
+
+public abstract class PendingEfCoreMigrationsChecker<TDbContext> : PendingMigrationsCheckerBase
+    where TDbContext : DbContext
+{
+    protected IUnitOfWorkManager UnitOfWorkManager { get; }
+    protected IServiceProvider ServiceProvider { get; }
+    protected ICurrentTenant CurrentTenant { get; }
+    protected IDistributedEventBus DistributedEventBus { get; }
+    protected string DatabaseName { get; }
+
+    protected PendingEfCoreMigrationsChecker(
+        IUnitOfWorkManager unitOfWorkManager,
+        IServiceProvider serviceProvider,
+        ICurrentTenant currentTenant,
+        IDistributedEventBus distributedEventBus,
+        string databaseName)
+    {
+        UnitOfWorkManager = unitOfWorkManager;
+        ServiceProvider = serviceProvider;
+        CurrentTenant = currentTenant;
+        DistributedEventBus = distributedEventBus;
+        DatabaseName = databaseName;
+    }
+
+    public virtual async Task<bool> CheckAsync()
+    {
+        var isMigrationRequired = false;
+
+        using (CurrentTenant.Change(null))
+        {
+            // Create database tables if needed
+            using (var uow = UnitOfWorkManager.Begin(requiresNew: true, isTransactional: false))
+            {
+                var pendingMigrations = await ServiceProvider
+                    .GetRequiredService<TDbContext>()
+                    .Database
+                    .GetPendingMigrationsAsync();
+
+                if (pendingMigrations.Any())
+                {
+                    await DistributedEventBus.PublishAsync(
+                        new ApplyDatabaseMigrationsEto
+                        {
+                            DatabaseName = DatabaseName
+                        }
+                    );
+                    isMigrationRequired = true;
+                }
+
+                await uow.CompleteAsync();
+            }
+
+            return isMigrationRequired;
+        }
+    }
+}
